Support for sending usage reports as email attachments
The generated usage report may now also be sent as an email attachment.
Configure an SMTP mailserver, subject, filename, recipients and a
suitable message in a YAML file and provide it to the report command
using the '--email' flag. Subject, filename and body supports Jinja2
templating to include the from- and to-dates covered in the report.

 zaptec-reporter -u myuser -p letmein report --email email_config.yml...